DeploymentManager/requirements.txt
azure-common==1.1.28
azure-core==1.26.3
azure-identity==1.12.0
azure-mgmt-compute==29.1.0
azure-mgmt-core==1.3.2
azure-storage-blob==12.15.0
bcrypt==4.0.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.2
docopt==0.6.2
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
msal==1.21.0
msal-extensions==1.0.0
msrest==0.7.1
numpy==1.24.2
oauthlib==3.2.2
pandas==1.5.3
pipreqs==0.4.11
portalocker==2.7.0
psutil==5.9.4
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
Python-EasyConfig==0.1.7
python-engineio==4.4.0
python-socketio==5.8.0
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
Resource==0.2.1
SQLAlchemy==2.0.9
urllib3==1.26.15
Werkzeug==2.2.3
